requirements.txt
PyQt5
matplotlib
opencv-python
scikit-image
numpy
ImageHash
Pillow
torch
transformers
piq